Enable Vulkan HPP in a crappy sample.
This code is lifted directly from RenderToy so we're over-initializing a lot of things, including the descriptor pools. We're not doing anything particularly interesting here except to create a presentation target and copy a Vulkan image into it. This should be the majority of the boilerplate for a simple Vulkan sample. We can't send this to master until the build can survive without Vulkan SDK installed.