Start to clean up D3D12 root signatures.
Small and boring looking change but this is actually quite significant - you don't need a root signature if you're just clearing a render target. There is no such thing as a "standard root signature" so we shouldn't pretend that there is. From here we're going to decouple the root signature from the device and move it into an encapsulated object.