Move DXR helpers into their own utility library.
DXR can be a bit ugly to work with at times just because it mixes CPU and GPU operation with uploads and syncs between phases. If we were writing something more sophisticated we'd be using call-with-continuation and spreading the CPU work over cores - we're not going that far here for clarity.