Use a common function for TLAS generation.
Remember: We aren't trying to be an amazingly efficient multithreading monster; we just want to build a TLAS in sync for demo. The process of building a TLAS doesn't require more than a big array of D3D12_RAYTRACING_INSTANCE_DESC structures so we can wrap the whole thing up here.