Start improving signature and heap creation.
We're using a mixin class to attach signatures and heaps to samples where we don't really reuse them that precisely. It's better if the samples just call into a helper function to build the standard signatures and heaps so each sample can be more precise about what it really uses.