Make matrix ordering consistent and add transforms to the DXR mesh sample.
This stuff is the bane of any graphic programmer. In general we all like to see row-major for D3D and column-major for OGL. That's not based on anything more than someones favorite breakfast cereal. For now we'll align all the code with row-major standard in order to avoid some confusing Transpose() operations.