Clang-format the universe!
We use the unmodified LLVM format since it's the most common in the wild (and the default). No sense adding configuration where none is required.

Philosophy: Any standard is better than no standard. The role of a clang-format is not to make the code how *you* like it, the purpose is to make the code how most *other* people like it...and there are more of them than there are of you.

"The needs of the many outweigh...the needs of the few" - Spock