Use Microsoft style interface naming and clean up.
Okay. Why? Easy - there are a lot of junior programmers that don't understand the concepts behind programming-to-contract, and these are the same programmers that will gladly add data members to your interface classes and turn them all in the glorified mixins. Later you'll find all your interfaces are glued together and your code looks like a tangled ball of garbage. Welcome to software.