Remove SceneCollector.
SceneCollector is great for prebuilding assets for the render but it suffers horribly once we get to anything more complex. Materials and textures just won't behave well in this system.

Instead we introduce the MutableMap: a particularly nasty piece of engineering which allows us to connect a factory to a map which will cache objects requested via a key. I won't dispute this is really nasty when it comes to scene updates but we're not at that point.