Just compute the TBN in the pixel shader.
Using vertex tangents/bitangents is a bit 90s - we can save a lot of tooling if we just avoid it entirely. Use the ShaderX formulation for the pixel-level TBN computed via partial derivatives.

There may be times when vertex level tangents are appropriate, but it requires a deep understanding of how tooling has exported maps and how attribute interpolation may be distorting the space. It's also far less friendly for skinning.