Eliminate all vertex tangent code.
Vertex tangents suck. With all the UV discontinuity problems, mesh fracturing, welding, and orthonormalizing, these things suck. They suck to process, they suck to calculate, and they just suck. Do everything in the pixel shader instead.

Note: If you were building a proper engine you'd want to optimize this and try to do more in the vertex shader. The only thing we care about is demonstrating the effect as directly as possible.