Get rid of ITransformSource.
This interface was added to throw data over the wall for UAV samples on OpenVR, but it's pretty nasty. Now we have a bona fide sample resource class we can just initialize one of those for the frame render. This should also allow us to build the left and right frame command buffers concurrently.