Get rid of ShaderToy.
Well, it was fun. The big problem with this demo is that it's showing how to reuse features in the context of a Win32 application. We actually don't want to expose anyone to Win32 if we can help it. If we're going to write a shader editing toy then we'll have to render our own UI to remain portable.