Introduce SBT helper.
SBT construction is a bit of a dark art. The data contained within this structure is built out of opaque identifiers and root parameters which must fall within specific locations of a flat byte buffer. Suffice to say, this isn't intuitive at all. Introduce an object to help map the SBT, dimension it, and locate various concepts and datum.