No need to persist the TLAS.
These statements are confusing so we'll make a good attempt to reduce them. When referencing an acceleration structure we often locate the structure in two buffers; TLAS and BLAS. The TLAS will usually have a reference within the descriptor heap and will be preserved by capture into the render lambda. The BLAS is only referenced indirectly from the TLAS so we have to deliberately persist it via some other mechanism.