Wildcard cs and dxr shaders.
If we see a .dxr.hlsl file then compile it with our common DXIL settings. Same concept for cs_5_0. Compiling offline gets to errors faster and speeds startup. It also means we can share library code with DXIL. We also set optimization to debug mode because we always end up doing that anyway.